// src/types/api.ts
export interface components {
    schemas: {
        CreateUserDto: Record<string, never>;
        UpdateUserDto: Record<string, never>;
        CreateBlogPostDto: Record<string, never>;
        UpdateBlogPostDto: Record<string, never>;
        /** @enum {string} */
        EBlogPostStatus: "ACTIVE" | "HIDDEN" | "DRAFT";
        UpdateBlogStatusDto: {
            /** @example ACTIVE */
            status: components["schemas"]["EBlogPostStatus"];
        };
        CreateBlockDto: Record<string, never>;
        UpdateBlockDto: Record<string, never>;
        CreateCommentDto: Record<string, never>;
        UpdateCommentDto: Record<string, never>;
        CreateHashtagDto: Record<string, never>;
        UpdateHashtagDto: Record<string, never>;
        CreateUserReactDto: Record<string, never>;
        UpdateUserReactDto: Record<string, never>;
        CreateEmojiDto: Record<string, never>;
        UpdateEmojiDto: Record<string, never>;
        CreateCommunityDto: Record<string, never>;
        UpdateCommunityDto: Record<string, never>;
        CreateReportDto: Record<string, never>;
        UpdateReportDto: Record<string, never>;
        CreateNotificationDto: Record<string, never>;
        UpdateNotificationDto: Record<string, never>;
        CreateSavedPostListDto: Record<string, never>;
        UpdateSavedPostListDto: Record<string, never>;
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}